requirements.txt
transformers==4.44.1
bitsandbytes
accelerate
torch
numpy
pandas
datasets
peft
scikit-learn
hydra-core
omegaconf
tf-keras